cleanup imports and contexts
