+++ programs/context.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../')))
+from src.context import *
