+++ programs/diamond_util.py
