+++ programs/prog_A.py
@@ -1,0 +1,10 @@
+from context import *
+
+ch = Channel[Offer[Send[str, Recv[int, End]], Send[int, End]]](('localhost', 5006), ('localhost', 5011))
+
+ch.send(1)
+ch.send({'str', 42})
+ch.send(42)
+
+for i in range(20):
+    ch.send(i)
