+++ tests/tests_sessiontypes/context.py
@@ -2,5 +2,4 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
 
-from src.context import *
-
+from tests.context import *
