+++ tests/tests_typechecking/test_typecheck.py
@@ -1,4 +1,5 @@
 from context import *
+from typing import List
 
 
 def get_func_return_type_in_latest_scope(tc, f) -> type:
